backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.59.6
motor==3.6.0
pymongo==4.10.1
python-multipart==0.0.19
python-dotenv==1.0.1
redis==5.2.1
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
Pillow==11.1.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
sse-starlette==2.2.1
